+++ src/DataContainer/PwaPushNotificationContainer.php
@@ -15,13 +15,11 @@ namespace HeimrichHannot\ContaoPwaBundle\DataContainer;
 use Contao\Config;
 use Contao\Controller;
 use Contao\Date;
-use Contao\News;
 use Contao\NewsModel;
 use Contao\PageModel;
-use Contao\System;
 use HeimrichHannot\ContaoPwaBundle\Model\PwaPushNotificationsModel;
-use Symfony\Component\Translation\TranslatorInterface;
 use Symfony\Component\Validator\Constraints\Url;
+use Symfony\Contracts\Translation\TranslatorInterface;
 
 class PwaPushNotificationContainer
 {